feat/1 useChatacter hook with CharactersContext
