+++ src/App.tsx
@@ -5,20 +5,23 @@ import { BrowserRouter, Routes, Route } from "react-router-dom";
 import FavoritesPage from "./pages/FavoritesPage";
 import CharactersListPage from "./pages/CharactersListPage";
 import CharacterDetailPage from "./pages/CharacterDetailPage";
+import { CharactersProvider } from "./hooks/useCharacters";
 
 function App() {
   return (
-    <BrowserRouter>
-      <AppLayout>
-        <Routes>
-          <Route path="/" element={<CharactersListPage />} />
-          <Route path="/characters" element={<CharactersListPage />} />
-          <Route path="/favorites" element={<FavoritesPage />} />
-          <Route path="/characters/:id" element={<CharacterDetailPage />} />
-          <Route path="/favorites/:id" element={<CharacterDetailPage />} />
-        </Routes>
-      </AppLayout>
-    </BrowserRouter>
+    <CharactersProvider>
+      <BrowserRouter>
+        <AppLayout>
+          <Routes>
+            <Route path="/" element={<CharactersListPage />} />
+            <Route path="/characters" element={<CharactersListPage />} />
+            <Route path="/favorites" element={<FavoritesPage />} />
+            <Route path="/characters/:id" element={<CharacterDetailPage />} />
+            <Route path="/favorites/:id" element={<CharacterDetailPage />} />
+          </Routes>
+        </AppLayout>
+      </BrowserRouter>
+    </CharactersProvider>
   );
 }
