+++ src/pages/CharactersListPage.tsx
@@ -1,20 +1,8 @@
-import { useEffect, useState } from "react";
-import * as rickAndMortyApi from "../services/api";
 import CharacterList from "../components/CharacterList";
-import type { Character } from "../types";
+import useCharacters from "../hooks/useCharacters";
 
 const CharactersListPage = () => {
-  const [characters, setCharacters] = useState<Character[] | null>(null);
-  useEffect(() => {
-    const fetchCharacters = async () => {
-      const { results: characters } = await rickAndMortyApi.getAllCharacters();
-      setCharacters(characters);
-    };
-
-    if (characters === null) {
-      fetchCharacters();
-    }
-  }, [characters]);
+  const { characters } = useCharacters();
   return (
     <div>
       {" "}
